Import plugins instead of require them
