+++ packages/gatsby/src/schema/graphql-engine/entry.ts
@@ -11,7 +11,7 @@ import { actions } from "../../redux/actions"
 import reporter from "gatsby-cli/lib/reporter"
 import { GraphQLRunner, IQueryOptions } from "../../query/graphql-runner"
 import { waitJobsByRequest } from "../../utils/wait-until-jobs-complete"
-import { setGatsbyPluginCache } from "../../utils/require-gatsby-plugin"
+import { setGatsbyPluginCache } from "../../utils/import-gatsby-plugin"
 import apiRunnerNode from "../../utils/api-runner-node"
 import type { IGatsbyPage, IGatsbyState } from "../../redux/types"
 import { findPageByPath } from "../../utils/find-page-by-path"
